del无用的功能
