--- examples/postman_test.json
@@ -1,83 +1,0 @@
-{
-	"info": {
-		"name": "Nahsor",
-		"_postman_id": "d20882de-6d71-84a3-6974-bbc70280c12d",
-		"description": "",
-		"schema": "https://schema.getpostman.com/json/collection/v2.0.0/collection.json"
-	},
-	"item": [
-		{
-			"name": "http://127.0.0.1:8808/querytcass",
-			"request": {
-				"method": "POST",
-				"header": [
-					{
-						"key": "Content-Type",
-						"value": "application/json"
-					}
-				],
-				"body": {
-					"mode": "raw",
-					"raw": "{\"idlist\":\"1,2,3\"}"
-				},
-				"url": "http://127.0.0.1:8808/querytcass"
-			},
-			"response": []
-		},
-		{
-			"name": "http://127.0.0.1:8808/addtcass",
-			"request": {
-				"method": "POST",
-				"header": [
-					{
-						"key": "Content-Type",
-						"value": "application/json"
-					}
-				],
-				"body": {
-					"mode": "raw",
-					"raw": "{\n \"moduleid\":8,\"explain\":\"测试自动导入功能2\",\"extract\":\"[]\",\"leader\":\"test\",\"remark\":\"备注嘻嘻\",   \"request\": \"{\\\"url\\\": \\\"http://127.0.0.1:2333/test\\\", \\\"json\\\": {\\\"aaa\\\": \\\"bbb\\\"}, \\\"method\\\": \\\"POST\\\", \\\"headers\\\": {\\\"Content-Type\\\": \\\"application/json\\\"}, \\\"timeout\\\": 10}\",\n    \"testname\": \"用例99\",\n    \"testtype\": \"testcass\",\n    \"validate\": \"[{\\\"Equal\\\": [\\\"r.json()\\\",\\\"request[\\\\\\\\\\\"json\\\\\\\\\\\"]\\\"]},{\\\"Equal\\\": [\\\"r.status_code\\\",\\\"200\\\"]}]\"\n}"
-				},
-				"url": "http://127.0.0.1:8808/addtcass"
-			},
-			"response": []
-		},
-		{
-			"name": "http://127.0.0.1:8808/addtcass",
-			"request": {
-				"method": "POST",
-				"header": [
-					{
-						"key": "Content-Type",
-						"value": "application/json"
-					}
-				],
-				"body": {
-					"mode": "raw",
-					"raw": "{\n  \"moduleid\":8,\"explain\":\"测试自动导入功能\",\"extract\":\"[]\",\"leader\":\"test\",\"remark\":\"备注\",  \"request\": \"{\\\"url\\\": \\\"http://127.0.0.1:2333/test\\\", \\\"json\\\": {\\\"aaa\\\": \\\"bbb\\\"}, \\\"method\\\": \\\"POST\\\", \\\"headers\\\": {\\\"Content-Type\\\": \\\"application/json\\\"}, \\\"timeout\\\": 10}\",\n    \"testname\": \"用例99\",\n    \"testtype\": \"testcass\",\n    \"validate\": \"[{\\\"Equal\\\": [\\\"r.json()\\\",\\\"request[\\\\\\\\\\\"json\\\\\\\\\\\"]\\\"]},{\\\"Equal\\\": [\\\"r.status_code\\\",\\\"200\\\"]}]\"\n}"
-				},
-				"url": "http://127.0.0.1:8808/addtcass"
-			},
-			"response": []
-		},
-		{
-			"name": "http://127.0.0.1:8808/testgo",
-			"request": {
-				"method": "POST",
-				"header": [
-					{
-						"key": "Content-Type",
-						"value": "application/json"
-					}
-				],
-				"body": {
-					"mode": "raw",
-					"raw": "{\"idlist\":\"42\"}"
-				},
-				"url": "http://127.0.0.1:8808/testgo",
-				"description": ""
-			},
-			"response": []
-		}
-	]
-}
