+++ nahsor/common.py
